+++ crates/teloxide-core/src/types/sticker_set.rs
@@ -2,7 +2,7 @@ use std::ops::Deref;
 
 use serde::{Deserialize, Serialize};
 
-use crate::types::{PhotoSize, Sticker, StickerFormat, StickerType};
+use crate::types::{PhotoSize, Sticker, StickerType};
 
 /// This object represents a sticker set.
 ///
@@ -19,10 +19,6 @@ pub struct StickerSet {
     /// Sticker type shared by all stickers in this set.
     #[serde(flatten)]
     pub kind: StickerType,
-
-    /// Sticker format shared by all stickers in this set.
-    #[serde(flatten)]
-    pub format: StickerFormat,
 
     /// List of all set stickers.
     pub stickers: Vec<Sticker>,
@@ -49,40 +45,6 @@ impl Deref for StickerSet {
     }
 }
 
-impl StickerSet {
-    /// Returns `true` is this is a "normal" raster sticker.
-    ///
-    /// Alias to [`self.format.is_raster()`].
-    ///
-    /// [`self.format.is_raster()`]: StickerFormat::is_raster
-    #[must_use]
-    pub fn is_raster(&self) -> bool {
-        self.format.is_raster()
-    }
-
-    /// Returns `true` is this is an [animated] sticker.
-    ///
-    /// Alias to [`self.format.is_animated()`].
-    ///
-    /// [`self.format.is_animated()`]: StickerFormat::is_animated
-    /// [animated]: https://telegram.org/blog/animated-stickers
-    #[must_use]
-    pub fn is_animated(&self) -> bool {
-        self.format.is_animated()
-    }
-
-    /// Returns `true` is this is a [video] sticker.
-    ///
-    /// Alias to [`self.format.is_video()`].
-    ///
-    /// [`self.format.is_video()`]: StickerFormat::is_video
-    /// [video]: https://telegram.org/blog/video-stickers-better-reactions
-    #[must_use]
-    pub fn is_video(&self) -> bool {
-        self.format.is_video()
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use crate::types::StickerSet;
@@ -93,8 +55,6 @@ mod tests {
         let json = r#"{
             "name": "teloxide_test",
             "title": "teloxide-test",
-            "is_animated": false,
-            "is_video": false,
             "sticker_type": "regular",
             "contains_masks": false,
             "stickers": [
@@ -106,6 +66,13 @@ mod tests {
                     "is_animated": false,
                     "is_video": false,
                     "type": "regular",
+                    "thumbnail": {
+                        "file_id": "AAMCAQADFQABYzB4ATH0sqXx351gZ5GpY1Z3Tl8AAlgCAAJ1t4hFbxNCoAg1-akBAAdtAAMpBA",
+                        "file_unique_id": "AQADWAIAAnW3iEVy",
+                        "file_size": 7698,
+                        "width": 320,
+                        "height": 320
+                    },
                     "thumb": {
                         "file_id": "AAMCAQADFQABYzB4ATH0sqXx351gZ5GpY1Z3Tl8AAlgCAAJ1t4hFbxNCoAg1-akBAAdtAAMpBA",
                         "file_unique_id": "AQADWAIAAnW3iEVy",
@@ -125,6 +92,13 @@ mod tests {
                     "is_animated": false,
                     "is_video": false,
                     "type": "regular",
+                    "thumbnail": {
+                        "file_id": "AAMCAQADFQABYzB4AcABR8-MuvGagis9Pk6liSAAAs8DAAL2YYBFNbvduoN1p7oBAAdtAAMpBA",
+                        "file_unique_id": "AQADzwMAAvZhgEVy",
+                        "file_size": 7780,
+                        "width": 320,
+                        "height": 320
+                    },
                     "thumb": {
                         "file_id": "AAMCAQADFQABYzB4AcABR8-MuvGagis9Pk6liSAAAs8DAAL2YYBFNbvduoN1p7oBAAdtAAMpBA",
                         "file_unique_id": "AQADzwMAAvZhgEVy",
@@ -141,7 +115,6 @@ mod tests {
 
         let set: StickerSet = serde_json::from_str(json).unwrap();
 
-        assert!(set.is_raster());
         assert!(set.is_regular());
         assert!(set.thumb.is_none());
         assert_eq!(set.stickers.len(), 2);